update db record after parsing config yaml and build log
1. update db record after parsing config yaml and build log
2. moved config and build log parsing to their own packages so that the sql package can import them
