+++ tools/monitoring/config/sample.yaml
