+++ tools/monitoring/log_parser/log_parser.go
@@ -14,35 +14,19 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package log_parser
 
 import (
 	"log"
 	"regexp"
+
+	"github.com/knative/test-infra/tools/monitoring/config"
 )
 
 // ErrorLog stores the error pattern and the corresponding error message found in the log
 type ErrorLog struct {
 	Pattern string
 	Msg     string
-}
-
-// compilePatterns compiles the patterns from string to Regexp. In addition it returns the list of
-// patterns that cannot be compiled
-func compilePatterns(patterns []string) ([]regexp.Regexp, []string) {
-	var regexps []regexp.Regexp
-	var badPatterns []string // patterns that cannot be compiled into regex
-
-	for _, pattern := range patterns {
-		r, err := regexp.Compile(pattern)
-		if err != nil {
-			log.Printf("Error compiling pattern [%s]: %v", pattern, err)
-			badPatterns = append(badPatterns, pattern)
-		} else {
-			regexps = append(regexps, *r)
-		}
-	}
-	return regexps, badPatterns
 }
 
 // collectMatches collects error messages that matches the patterns from text.
@@ -63,11 +47,11 @@ func collectMatches(regexps []regexp.Regexp, text []byte) []ErrorLog {
 // ParseLog fetches build log from given url and checks it against given error patterns. Return
 // all found error patterns and error messages in pairs.
 func ParseLog(url string, patterns []string) ([]ErrorLog, error) {
-	content, err := getFileBytes(url)
+	content, err := config.GetFileBytes(url)
 	if err != nil {
 		return nil, err
 	}
-	regexps, badPatterns := compilePatterns(patterns)
+	regexps, badPatterns := config.CompilePatterns(patterns)
 	if len(badPatterns) != 0 {
 		log.Printf("The following patterns cannot be compiled: %v", badPatterns)
 	}
